requirements.txt
Pillow>=7.2.0
alive-progress>=1.6.1
SSIM-PIL>=1.0.10
PyQt5>=5.15.1